- refactor : structure code
