+++ tasks-app/src/apollo/config.ts
@@ -1,0 +1,16 @@
+import {
+    ApolloClient,
+    createHttpLink,
+    InMemoryCache
+} from '@apollo/client';
+
+const httpLink = createHttpLink({
+    uri: 'http://localhost:8000/graphql/'
+});
+
+const apolloClient = new ApolloClient({
+    link: httpLink,
+    cache: new InMemoryCache()
+});
+
+export default apolloClient;
