+++ tasks-app/src/index.tsx
@@ -2,26 +2,12 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 import { BrowserRouter } from 'react-router-dom';
-
-import {
-    ApolloProvider,
-    ApolloClient,
-    createHttpLink,
-    InMemoryCache
-} from '@apollo/client';
-
-const httpLink = createHttpLink({
-    uri: 'http://localhost:8000/graphql/'
-});
-
-const client = new ApolloClient({
-    link: httpLink,
-    cache: new InMemoryCache()
-});
+import { ApolloProvider } from '@apollo/client';
+import apolloClient from "./apollo/config";
 
 ReactDOM.render(
     <BrowserRouter>
-        <ApolloProvider client={client}>
+        <ApolloProvider client={apolloClient}>
             <App />
         </ApolloProvider>
     </BrowserRouter>,
